+++ client-src/main.js
@@ -10,7 +10,8 @@ const spotifyDemos = [
 			"Back-end: Node.js with Express",
 			"Designed a REST API for authenticating, getting liked songs, and setting and clearing song ranges",
 			"Emphasized architecture in the back-end by separating the REST API, HTTP requests to Spotify, utility functions, and commind-line argument parsing into separate modules.",
-			"Utilized toasts (Toastify) to notify the user of successful changes"
+			"Utilized toasts (Toastify) to notify the user of successful changes",
+			"Supported local development through command-line arguments"
 		]
 	},
 	{
@@ -25,7 +26,8 @@ const spotifyDemos = [
 			"Developed my own recursive algorithm for placing squares of random size in a larger square",
 			"Utilized local storage to track user authentication between sessions",
 			"Converted code from JavaScript to TypeScript to allow for stronger typing and abstraction",
-			"Utilized the WebSocket protocol for communication between server and client"
+			"Utilized the WebSocket protocol for communication between server and client",
+			"Supported local development through command-line arguments"
 		]
 	},
 	{
@@ -39,7 +41,8 @@ const spotifyDemos = [
 			"Back-end: Node.js",
 			"First project interfacing with Spotify",
 			"Leveraged Chart.js to allow analysis of a user's Liked Songs",
-			"Utilized WebSockets to facilitate back-and-forth communicating of server and client"
+			"Utilized WebSockets to facilitate back-and-forth communicating of server and client",
+			"Supported local development through command-line arguments"
 		]
 	},
 ];
@@ -54,7 +57,7 @@ const sections = [
 		"projects": spotifyDemos
 	},
 	{
-		"title": "Other Projects",
+		"title": "Other Personal Projects",
 		"projects": [
 			{
 				"title": "Spotify Console",
@@ -75,7 +78,8 @@ const sections = [
 				"points": [
 					"What you're looking at right now!",
 					"Implemented a reverse proxy system for running demos (each demo is locally running on a separate port)",
-					"Emphasized semantic versioning and detailed changelogs (refer to GitHub repo)"
+					"Emphasized semantic versioning and detailed changelogs (refer to GitHub repo)",
+					"Standardized properties of 'sections' (e.g. 'Spotify Demos', 'Other Personal Projects') and 'projects' (e.g. 'Playlist Art', 'Adventurous Art') to allow for polymorphism in rendering content"
 				]
 			},
 			{
@@ -89,11 +93,47 @@ const sections = [
 					"Implemented real-time collaborative drawing using WebSockets",
 					"Incorporated the Met API to allow users to trace over artwork (picture 4)",
 					"Designed a large portion of general UI look and logic",
+					"Configured CI/CD with unit tests running on every branch and an APK being deployed on master",
 					"Worked in an Agile environment"
 				]
 			}
 		]
 	},
+	{
+		"title": "Professional Experience",
+		"projects": [
+			{
+				"title": "Spreetail - Software Engineering Intern",
+				"points": [
+					"Location: Omaha, Nebraska",
+					"May 2019 - August 2019",
+					"Fully engaged in Agile methodology, including working within 2-week sprints, writing discovery tickets, and attending and contributing to standup, retro, grooming, and sprint planning meetings",
+					"Programmed using JavaScript with React and Redux on the frontend and C# with .NET and LINQ on the backend",
+					"Created and worked on an entirely new project to be used internally, which included setting up build tools for its npm package, setting up a testing framework, and designing the overall architecture of the project",
+					"Statistics: 25 JIRA tickets completed, 45 merge requests merged across 6 repositories, 135 discussions resolved within merge requests, and 736 total contributions"
+				]
+			},
+			{
+				"title": "IERO, LLC. - Freelance Software Developer",
+				"points": [
+					"Location: Ames, Iowa (remote)",
+					"February 2019 - May 2019",
+					"Collaborated with three other Iowa State students to build a prototype Android application in Java",
+					"Configured multiple AWS services, including Lambda, API Gateway, RDS (database), and VPC (security)",
+					"Created API specifications and architecture, navigation, and sequence diagrams to aid in the product’s design"
+				]
+			},
+			{
+				"title": "Workiva - Software Engineering Intern",
+				"points": [
+					"Collaborated with a full development team throughout the summer, including multiple programmers, a quality assurance lead, and a user experience lead",
+					"Contributed directly to a future product by completing 44 programming tickets in an Agile environment",
+					"Leveraged Git for version control, which included strict code and quality assurance reviews, resulting in 7,698 additions and 3,376 deletions",
+					"Presented to a group of executives (including CFO Stuart Miller) on a potential new product for a competition where I placed first out of fifteen groups"
+				]
+			}
+		]
+	}
 ];
 
 const ProjectLink = ({title, onClick}) => (