+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.4] - 2019-09-29
+### Added
+- "Professional Experience" section
+  - Spreetail, IERO, Workiva
+- New points added to existing projects
+
+### Changed
+- Header rename: "Other Projects" -> "Other Personal Projects"
+- Minor CSS change
+
 ## [1.2.3] - 2019-09-25
 ### Added
 - Added a message for when no project is selected